refactor : 기능별로 코드 분할
